+++ demisto_sdk/commands/common/errors.py
@@ -151,6 +151,7 @@ ERROR_CODE = {
     "pack_metadata_version_should_be_raised": "PA114",
     "pack_timestamp_field_not_in_iso_format": 'PA115',
     "invalid_package_dependencies": "PA116",
+    "pack_metadata_invalid_support_type": "PA117",
     "pack_metadata_certification_is_invalid": "PA118",
     "pack_metadata_non_approved_usecases": "PA119",
     "pack_metadata_non_approved_tags": "PA120",
@@ -949,6 +950,11 @@ class Errors:
     @error_code_decorator
     def pack_metadata_missing_url_and_email():
         return 'Contributed packs must include email or url.'
+
+    @staticmethod
+    @error_code_decorator
+    def pack_metadata_invalid_support_type(pack_meta_file):
+        return 'Support field should be one of the following: xsoar, partner, developer or community.'
 
     @staticmethod
     @error_code_decorator