+++ demisto_sdk/commands/generate_docs/tests/generate_docs_test.py
@@ -311,7 +311,6 @@ def test_get_input_data_complex():
 
     assert _value == 'File.Name'
 
-
 # script tests